proyecto

+++ PRO/Proyecto/backend/tuturno/tuturno/src/main/java/com/tuturno/dto/servicio/ServicioRequestDTO.java
@@ -1,10 +1,8 @@
 package com.tuturno.dto.servicio;
 
-import java.math.BigDecimal;
-
 public record ServicioRequestDTO(
         String nombre,
         String descripcion,
-        BigDecimal precio,
+        Double precio,
         Integer duracionMinutos // En el JSON vendrá como "duracionMinutos"
 ) {}
